requirements.txt
fastapi
uvicorn
watchfiles
sqlalchemy
passlib==1.7.4
bcrypt==3.2.0
python-jose
python-multipart
email-validator
dotenv